Implementado el Cron Task para calcular los grados de confianza (Domingos 13:00hs).
Refactor de integración con la API, cambié el tipo de dato del puntaje de int  -> Float.

Los test de integración pasan.

+++ main-project/src/main/java/ar/edu/utn/frba/dds/modelos/calculo_grado_confianza/UsuarioAEvaluar.java
@@ -8,19 +8,19 @@ import lombok.Setter;
 @Setter
 public class UsuarioAEvaluar {
   public Usuario usuario;
-  public int puntajeInicial;
+  public Float puntajeInicial;
 
   public UsuarioAEvaluar() {
     this.usuario = new Usuario();
-    this.puntajeInicial = 0;
+    this.puntajeInicial = 0f;
   }
 
   public UsuarioAEvaluar(Usuario usuario) {
     this.usuario = usuario;
-    this.puntajeInicial = 0;
+    this.puntajeInicial = 0f;
   }
 
-  public UsuarioAEvaluar(Usuario usuario, int puntajeInicial) {
+  public UsuarioAEvaluar(Usuario usuario, Float puntajeInicial) {
     this.usuario = usuario;
     this.puntajeInicial = puntajeInicial;
   }
