+++ main-project/src/main/java/ar/edu/utn/frba/dds/modelos/calculo_grado_confianza/servicio_externo/api_models/RequestUsuario.java
@@ -10,5 +10,5 @@ public class RequestUsuario {
   @SerializedName("id")
   public int id;
   @SerializedName("puntaje_inicial")
-  public int puntajeInicial;
+  public Float puntajeInicial;
 }
